extracted pipeline actions into its own component
